requirements.txt
streamlit
pandas
psycopg2-binary
python-docx
openpyxl
fpdf